fix loss plots
